+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Fix features rendering at the wrong position when a layer uses a `within` filter and the map uses a projection other than Mercator or globe.
 - Fix global placement throwing an error when placing symbols with no text.
 - Fix setting `padding` as a sideeffect in a `preloadOnly` animation.
+- Fix symbol fading not finishing when using global placement.
 
 ## 3.31.0